# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Flare by TK Restaurant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (admin — all; customer can track their own order via ID)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [new, preparing, ready, completed, cancelled]
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          description: Unauthorized
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order (guest)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Validation error
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order by ID (for customer order tracking)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          description: Unauthorized
        "404":
          description: Not found

  # ── Admin Dashboard ───────────────────────────────────────────
  /admin/dashboard:
    get:
      operationId: getAdminDashboard
      tags: [admin]
      summary: Get admin dashboard stats
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
        "401":
          description: Unauthorized
components:
  schemas:
...
    OrderItem:
      type: object
      required: [menuItemId, name, price, quantity]
      properties:
        menuItemId:
          type: integer
        name:
          type: string
        price:
          type: number
        quantity:
          type: integer
    Order:
      type: object
      required: [id, customerName, customerPhone, orderType, status, totalAmount, items, createdAt]
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        customerAddress:
          type: ["string", "null"]
        orderType:
          type: string
          enum: [delivery, pickup]
        status:
          type: string
          enum: [new, preparing, ready, completed, cancelled]
        totalAmount:
          type: number
        specialInstructions:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
    OrderInput:
      type: object
      required: [customerName, customerPhone, orderType, items]
      properties:
        customerName:
          type: string
          minLength: 1
        customerPhone:
          type: string
          minLength: 1
        customerAddress:
          type: string
        orderType:
          type: string
          enum: [delivery, pickup]
        specialInstructions:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
    OrderItemInput:
      type: object
      required: [menuItemId, quantity]
      properties:
        menuItemId:
          type: integer
        quantity:
          type: integer
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [new, preparing, ready, completed, cancelled]

    # ── Dashboard ──────────────────────────────────────────
    DashboardStats:
      type: object
      required: [totalOrders, todayOrders, totalRevenue, todayRevenue, ordersByStatus, recentOrders]
      properties:
        totalOrders:
          type: integer
        todayOrders:
          type: integer
        totalRevenue:
          type: number
        todayRevenue:
          type: number
        ordersByStatus:
          type: object
          properties:
            new:
              type: integer
            preparing:
              type: integer
            ready:
              type: integer
            completed:
              type: integer
            cancelled:
              type: integer
        recentOrders:
          type: array
          items:
            $ref: "#/components/schemas/Order"