# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Flare by TK Restaurant API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderOrderType } from './orderOrderType';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  customerName: string;
  customerPhone: string;
  /** @nullable */
  customerAddress?: string | null;
  orderType: OrderOrderType;
  status: OrderStatus;
  totalAmount: number;
  /** @nullable */
  specialInstructions?: string | null;
  items: OrderItem[];
  createdAt: string;
}