# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Flare by TK Restaurant API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInputOrderType } from './orderInputOrderType';
import type { OrderItemInput } from './orderItemInput';

export interface OrderInput {
  /** @minLength 1 */
  customerName: string;
  /** @minLength 1 */
  customerPhone: string;
  customerAddress?: string;
  orderType: OrderInputOrderType;
  specialInstructions?: string;
  items: OrderItemInput[];
}