# lib/api-zod/src/generated/types/orderItem.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Flare by TK Restaurant API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItem {
  menuItemId: number;
  name: string;
  price: number;
  quantity: number;
}